refactor: 서버 컴포넌트를 위해 메인 페이지 바텀 섹션 분리
