+++ src/app/page.tsx
@@ -1,12 +1,6 @@
 'use client';
 
-import { useMediaQuery, useTheme } from '@mui/material';
 import Box from '@mui/material/Box';
-import Button from '@mui/material/Button';
-import Typography from '@mui/material/Typography';
-import Image from 'next/image';
-import { Autoplay, Pagination } from 'swiper/modules';
-import { Swiper, SwiperSlide } from 'swiper/react';
 import { useGetStore } from '@/api/store';
 import HideScrollTop from '@/components/common/hide-scroll-top/HideScrollTop';
 import { SectionTitle } from '@/components/common/section-title';
@@ -14,10 +8,10 @@ import NewlyAddedStore from '@/components/main/newly-added-store/NewlyAddedStore
 import RecommendedStore from '@/components/main/recommended-store/RecommendedStore';
 import { SectionLayout } from '@/components/main/section-layout';
 import SuccessExample from '@/components/main/success-example-store/SuccessExampleStore';
+import { MainHero } from '@/components/main/main-hero';
+import { BottomSection } from '@/components/main/bottom-section';
 
 export default function Home() {
-  const theme = useTheme();
-  const isDownMedium = useMediaQuery(theme.breakpoints.down('md'));
   const { data, isLoading } = useGetStore({ page: '1', limit: '20' });
 
   return (
@@ -31,106 +25,14 @@ export default function Home() {
         paddingBottom: 10,
       }}
     >
-      <Box
-        sx={{
-          '& .swiper': {
-            width: '100%',
-            height: '400px',
-          },
-        }}
-      >
-        <Swiper
-          modules={[Autoplay, Pagination]}
-          loop
-          grabCursor
-          pagination={{ clickable: true }}
-        >
-          <SwiperSlide>
-            <Box
-              sx={{
-                position: 'relative',
-                display: 'flex',
-                width: '100%',
-                height: '100%',
-                justifyContent: 'center',
-                alignItems: 'center',
-              }}
-            >
-              <Box
-                sx={{
-                  display: 'flex',
-                  flexDirection: 'column',
-                  justifyContent: 'center',
-                  alignItems: 'center',
-                  zIndex: 9,
-                  color: 'primary.contrastText',
-                }}
-              >
-                <Typography variant='h4'>코어창업</Typography>
-                <Typography>코어창업이</Typography>
-              </Box>
-              <Box
-                sx={{
-                  '&:after': {
-                    position: 'absolute',
-                    top: 0,
-                    right: 0,
-                    bottom: 0,
-                    left: 0,
-                    content: '""',
-                    background: 'rgba(0, 0, 0, 0.5)',
-                  },
-                }}
-              >
-                <Image fill src={'/logo.png'} alt='hero image' />
-              </Box>
-            </Box>
-          </SwiperSlide>
-          <SwiperSlide>
-            <Image fill src={'/logo.png'} alt='hero image' />
-          </SwiperSlide>
-        </Swiper>
-      </Box>
+      <MainHero />
       <RecommendedStore storeDataArr={data.data} isLoading={isLoading} />
       <SuccessExample />
       <NewlyAddedStore storeDataArr={data.data} isLoading={isLoading} />
       <SectionLayout color='white'>
         <SectionTitle label='협업 브랜드' />
       </SectionLayout>
-      <SectionLayout color='white' innerSx={{ gap: 8 }}>
-        {isDownMedium ? (
-          <Box
-            sx={{
-              display: 'flex',
-              flexDirection: 'column',
-              alignItems: 'center',
-              gap: 2,
-            }}
-          >
-            <Typography variant='h4' component='h3' fontWeight='bold'>
-              창업의 문을 여는 열쇠
-            </Typography>
-            <Typography
-              variant='h4'
-              component='h3'
-              fontWeight='bold'
-              color='primary.main'
-            >
-              코어창업
-            </Typography>
-          </Box>
-        ) : (
-          <Typography variant='h4' component='h3' fontWeight='bold'>
-            {'창업의 문을 여는 열쇠, '}
-            <Box display='inline' color='primary.main'>
-              코어창업
-            </Box>
-          </Typography>
-        )}
-        <Button variant='contained' size='large' sx={{ fontWeight: 'bold' }}>
-          상담 신청하기
-        </Button>
-      </SectionLayout>
+      <BottomSection />
       <HideScrollTop />
     </Box>
   );
