+++ src/components/main/bottom-section/index.ts
@@ -1,0 +1,1 @@
+export { default as BottomSection } from './BottomSection';
